workers/inventory-service/requirements.txt
aio_pika
asyncio
psycopg2-binary